fix(security): corrigindo erro 403 na interface de admin
- Ajusta permissões para criar/editar mesas, funcionários e itens

- Atualiza algumas configurações de segurança e inicialização de dados para gerar um usuario admin (apenas para dev, será removido em prod)

+++ backend/src/main/java/br/com/restaurante/gestao_restaurante/dto/barman/BarmanCreateDTO.java
@@ -10,6 +10,6 @@ public class BarmanCreateDTO {
     private String email;
     private String cpf;
     private String senha;
-    private BigDecimal salario;
+    private Double salario;
     private String matricula;
 }
